requirements.txt
amqp==5.1.1
billiard==4.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
kombu==5.3.1
pika==1.3.2
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6